DEVEXP-349: Run Highlighted Code
This includes one new feature and some refactoring.

Run Highlighted Code : This allows a user "eval" only the selected code in the current editor tab. This works for all the eval functions (XQY, JS, SQL, SPARQL, Rows).

Refactoring : Completely "class-ified" the vscQueryDirectory and renamed the file and class to editorQueryEvaluator.