DEVEXP-419: Show error when "Disconnect" can't get the info
Adding some tests for the "Disconnect" function.
This includes adding the "sinon" library for mocking objects & functions which allows us to test calls the the VSCode UI.

Lot's of renaming variables related to testing.
Gradual reordering of import statements in an effort to standardize.

Added a certificate template to the test-app. At this point it is only for manual testing/development of SSL features.