DEVEXP-420: Launching Eval fails with self-signed cert
This one requires manual testing.
The primary changes for this story:
* extension - register the new custom message handler
* xqyDebug - stops the request when launching with ML fails.
* xqyDebugManager - added a handler for custom events from the Debug Adapter, including the new one generated when launching with ML fails.
* xqyRuntime - when launching the query with ML fails, creates a new standard MLXPRS error object and returns that object to the debug client via a custom event.

I attempted to write an integration test for it, but was unable to get the dynamic Debug Client to work properly with the dynamic XQY Debug Adapter.
However, during that attempt, I did some refactoring to the integration test directory that will help with future efforts at XQY debugging integration testing.

Also:
Added some debug-level logging for interactions with the Debug Adapters.