DEVEXP-393 - Run marklogic-unit-test module
Adds the feature to run a single ML Unit Test - based on the current editor tab.

Added a new configure option to specify the port that the unit tests are running on.

Update README - including unrelated tweaks.

Moved a couple helper functions that are only used for the MLXPRS webview to the webview provider.

Includes a little refactoring of MarkLogic client generation in the TestHelper file.