MLE-24928 Fixing CVEs
Main one - bumping webpack to 5.102.1 .

Also bumping mocha to 10.8.2 to knock out a couple other CVEs.

Want to see what Black Duck thinks about this. Have not done any testing of this, will address that later when we plan to do a new release. At which point, we will likely update to Node Client 4.0 and require Node 22, and thus update types/node (which the webpackFix comment states is likely necessary).